Task #4796: Transport delays and phase offsets in double precision from casacore to the GPUs

+++ branches/Cobalt-Task4796/RTCP/Cobalt/CoInterface/src/SubbandMetaData.h
@@ -40,8 +40,8 @@ namespace LOFAR
       SubbandMetaData(size_t nrTABs = 0);
 
       struct beamInfo {
-        float delayAtBegin;
-        float delayAfterEnd;
+        double delayAtBegin;
+        double delayAfterEnd;
         double beamDirectionAtBegin[3];
         double beamDirectionAfterEnd[3];
       };
